requirements.txt
yt-dlp>=2023.3.4
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.0.0
ffmpeg-python>=0.2.0
python-dotenv>=0.19.0
pydantic>=2.0.0
typer>=0.9.0
rich>=10.0.0
pillow
pandas

# Development dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
isort>=5.10.0
mypy>=0.950
pylint>=2.15.0
pre-commit>=2.20.0
